refactor: remove packer from legacy
