+++ package.json
@@ -161,7 +161,6 @@
     "inquirer": "6.5.2",
     "inquirer-fuzzy-path": "2.3.0",
     "is-glob": "4.0.1",
-    "is-relative-path": "2.0.0",
     "is-url": "1.2.4",
     "isbinaryfile": "4.0.6",
     "lodash": "4.17.21",
@@ -211,7 +210,6 @@
     "semver": "7.5.2",
     "serialize-error": "5.0.0",
     "ssh2": "1.4.0",
-    "ssri": "10.0.1",
     "string-format": "0.5.0",
     "stringify-package": "1.0.1",
     "stylus-lookup": "3.0.2",
