--- src/pack/index.ts
@@ -1,2 +1,0 @@
-export { Packer } from './packer';
-export type { PackResultMetadata, PackWriteOptions, PackOptions, PackResult } from './packer';
