+++ workspace.jsonc
@@ -306,6 +306,7 @@
         "is-binary-path": "2.1.0",
         "is-builtin-module": "3.1.0",
         "is-primitive": "3.0.1",
+        "is-relative-path": "2.0.0",
         "jest": "27.5.1",
         "jest-message-util": "27.5.1",
         "jest-watcher": "27.5.1",
@@ -364,6 +365,7 @@
         "semver-range-intersect": "0.3.1",
         "socket.io-client": "2.4.0",
         "socks-proxy-agent": "5.0.0",
+        "ssri": "10.0.1",
         "strip-ansi": "6.0.0",
         "subscriptions-transport-ws": "0.9.18",
         "symlink-dir": "^5.1.1",
